pos y globalstiff corregidos y test de eficiencia implementados

+++ python/triangularElements/linel.py
@@ -4,7 +4,7 @@ def linel(struct):
     nelemy = struct["nelemy"][0]
     nelem = int(struct["nelem"])
     nsquare = int(nelem/2)
-    lin = zeros((6, nelem))
+    lin = zeros((6, nelem), dtype= int)
 
     for i in range(1, nsquare+1):
         node = i + int(i/(nelemy/2)) if i%(nelemy/2) != 0 else i + int(i/(nelemy/2))- 1
